+++ smx/sourcemod/printf.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 import dataclasses
 import re
 from ctypes import c_int
-from functools import wraps
-from typing import Sequence, TYPE_CHECKING
+from functools import reduce, wraps
+from typing import Iterable, Sequence, TYPE_CHECKING
 
 from smx.definitions import cell, ucell
 from smx.exceptions import SourcePawnStringFormatError
@@ -26,13 +26,13 @@ def _check_fmt_args(x, arg, args):
             (arg, args))
 
 
-def has_flag(v, flag):
+def has_flag(v: int, flag: int):
     return (v & flag) == flag
 
 
-def has_flags(v, flags):
-    # TODO: use generator to allow for early escape (profile first)
-    return all(map(lambda f: has_flag(v, f), flags))
+def has_flags(v: int, flags: Iterable[int]):
+    merged_flags = reduce(lambda a, b: a | b, flags, 0)
+    return has_flag(v, merged_flags)
 
 
 def atoi(s, length=False):
@@ -57,7 +57,7 @@ def render_fmt(fmt_char, flags, width, precision=None):
 def render_int(val, flags, width):
     s = render_fmt('d', flags, width) % val
     # SourceMod differs from Python in that it will left adjust with zeroes
-    if has_flags(flags, (FMT_LADJUST,FMT_ZEROPAD)):
+    if has_flags(flags, (FMT_LADJUST, FMT_ZEROPAD)):
         s = s.replace(' ', '0')
     return s
 
@@ -74,7 +74,7 @@ def render_bin(val, flags, width):
 def render_hex(val, flags, width):
     s = render_fmt('x', flags, width) % val
     # SourceMod differs from Python in that it will left adjust with zeroes
-    if has_flags(flags, (FMT_LADJUST,FMT_ZEROPAD)):
+    if has_flags(flags, (FMT_LADJUST, FMT_ZEROPAD)):
         s = s.replace(' ', '0')
     return s
 
@@ -84,22 +84,26 @@ def render_float(val, flags, width, precision):
     return render_fmt('f', flags, width, precision) % val
 
 
-def formatfunc(matches, eats=0, incr=True):
+def formatfunc(matches: str, eats: int = 0, incr: bool = True):
     """
     Marks a function as a format handler
-    @type   matches: iterable -> str[1]
-    @param  matches: An iterable of characters the format function supports
-    @type   eats: int
-    @param  eats: The number of arguments the format function uses up on
-                successful formatting (returning non-None)
-    @type   incr: bool
-    @param  incr: Whether or not to increment i after return
+
+    :param matches:
+        An iterable of characters the format function supports
+
+    :param eats:
+        The number of arguments the format function uses up on successful formatting
+        (i.e., returning non-None)
+
+    :param incr:
+        Whether to advance cursor after return
+
     """
     def inner(f):
         func = f
         if eats > 0:
             def wrapper(*args, **kwargs):
-                state = kwargs['state'] if 'state' in kwargs else args[-1]
+                state = kwargs.get('state', args[-1])
                 _check_fmt_args(eats, state.arg - 1, state.num_args)
                 return f(*args, **kwargs)
             func = wraps(f)(wrapper)
@@ -119,7 +123,7 @@ def isformatfunc(f):
 
 @dataclasses.dataclass
 class PrintfState:
-    i: int
+    cursor: int
     fmt: str
     out: str
     flags: int
@@ -148,10 +152,10 @@ class PrintfFormatter:
 
     @formatfunc('.', incr=False)
     def precision(self, ch: str, state: PrintfState):
-        state.i += 1  # Eat the period
-        prec, chars = atoi(state.fmt[state.i:], length=True)
+        state.cursor += 1  # Eat the period
+        prec, chars = atoi(state.fmt[state.cursor:], length=True)
         state.precision = None if prec < 0 else prec
-        state.i += chars
+        state.cursor += chars
         state.sz_format += '.' + str(prec)
 
     @formatfunc('0')
@@ -160,8 +164,8 @@ class PrintfFormatter:
 
     @formatfunc('123456789', incr=False)
     def width(self, ch: str, state: PrintfState):
-        state.width,chars = atoi(state.fmt[state.i:], length=True)
-        state.i += chars
+        state.width, chars = atoi(state.fmt[state.cursor:], length=True)
+        state.cursor += chars
         state.sz_format += str(state.width)
 
     @formatfunc('%')
@@ -170,9 +174,9 @@ class PrintfFormatter:
 
     @formatfunc('c', eats=1)
     def char(self, ch: str, state: PrintfState):
-        addr = state.amx._getheapcell(state.params[state.arg])
-        val = state.amx._local_to_char(addr)
-        return val[0]
+        addr = state.params[state.arg]
+        int_val = state.amx._getheapchar(addr)
+        return chr(int_val)
 
     @formatfunc('id', eats=1)
     def integer(self, ch: str, state: PrintfState):
@@ -199,7 +203,7 @@ class PrintfFormatter:
             val = '(null)'
             state.precision = None
         else:
-            val = state.amx._local_to_string(state.params[state.arg])
+            val = state.amx._getheapstring(state.params[state.arg])
         return render_string(val, state.flags, state.width, state.precision)
 
     @formatfunc('b', eats=1)
@@ -242,7 +246,7 @@ def atcprintf(
     fmt_len = len(fmt)
 
     state = PrintfState(
-        i=0,
+        cursor=0,
         fmt=fmt,
         out='',
         flags=0,
@@ -254,17 +258,17 @@ def atcprintf(
         params=params,
         sz_format=None,
     )
-    while state.i < fmt_len:
-        percent = state.fmt[state.i:].find('%')
+    while state.cursor < fmt_len:
+        percent = state.fmt[state.cursor:].find('%')
         if percent == -1:
-            state.out += state.fmt[state.i:]
+            state.out += state.fmt[state.cursor:]
             break
 
         else:
-            state.out += state.fmt[state.i:state.i+percent]
-            state.i += percent
+            state.out += state.fmt[state.cursor:state.cursor + percent]
+            state.cursor += percent
             # Skip the percent
-            state.i += 1
+            state.cursor += 1
 
         state.flags = 0
         state.width = 0
@@ -272,8 +276,8 @@ def atcprintf(
         state.sz_format = '%'
 
         do_break = False
-        while state.i < fmt_len:
-            ch = state.fmt[state.i]
+        while state.cursor < fmt_len:
+            ch = state.fmt[state.cursor]
             if ch in formatter.format_chars:
                 f = formatter.format_chars[ch]
                 out = f(ch, state)
@@ -285,7 +289,7 @@ def atcprintf(
 
                 if f.incr:
                     state.sz_format += ch
-                    state.i += 1
+                    state.cursor += 1
 
                 if do_break:
                     break